packages/backend/all_requirements.txt
annotated-types==0.6.0
anyio==4.2.0
arandr==0.1.11
astroid==3.0.2
attrs==23.1.0
autocommand==2.2.2
autopep8==2.0.4
black==23.12.1
btrfsutil==6.7
cattrs==23.2.3
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.7
dill==0.3.7
distlib==0.3.8
docstring-to-markdown==0.13
fastapi==0.109.0
fastjsonschema==2.19.1
filelock==3.13.1
flake8==7.0.0
greenlet==3.0.3
idna==3.6
importlib-metadata==7.0.1
inflect==7.0.0
iniconfig==2.0.0
isort==5.13.2
jaraco.context==4.3.0
jaraco.functools==3.9.0
jaraco.text==3.12.0
jedi==0.19.1
jedi-language-server==0.41.2
lsprotocol==2023.0.0
Mako==1.3.0
Markdown==3.5.1
MarkupSafe==2.1.4
mccabe==0.7.0
more-itertools==10.1.0
msgpack==1.0.7
mypy-extensions==1.0.0
ordered-set==4.1.0
packaging==23.2
parso==0.8.3
pathspec==0.12.1
platformdirs==4.1.0
pluggy==1.3.0
ply==3.11
psycopg2==2.9.9
pycairo==1.25.1
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pydocstyle==6.3.0
pyflakes==3.2.0
pygls==1.2.1
PyGObject==3.46.0
pylint==3.0.3
pynvim==0.5.0
pytest==7.4.4
python-dotenv==1.0.1
python-lsp-black==2.0.0
python-lsp-jsonrpc==1.1.2
python-lsp-server==1.9.0
pytoolconfig==1.3.1
QTermWidget==1.4.0
requests==2.31.0
rope==1.12.0
sniffio==1.3.0
snowballstemmer==2.2.0
SQLAlchemy==2.0.25
starlette==0.35.1
tomli==2.0.1
tomlkit==0.12.3
trove-classifiers==2024.1.9
typing_extensions==4.9.0
ujson==5.9.0
urllib3==2.1.0
validate-pyproject==0.13.post1.dev0+gb752273.d20230520
virtualenv==20.25.0
websockets==12.0
whatthepatch==1.0.5
yapf==0.40.2
zipp==3.17.0
